+++ tests/test__valid.py
@@ -199,10 +199,10 @@ class Test__ValidTypes:
     @pytest.mark.parametrize(
         argnames="source, args, _EXPECTED",
         argvalues=[
-            (1, (ValueNotPassed, lambda x: float(x) >= 1), True),
-            ("1", (ValueNotPassed, lambda x: float(x) >= 1), True),
-            ("0", (ValueNotPassed, lambda x: float(x) >= 1), False),
-            ("hello", (ValueNotPassed, lambda x: float(x) >= 1), False),
+            (1, (ValueNotExist, lambda x: float(x) >= 1), True),
+            ("1", (ValueNotExist, lambda x: float(x) >= 1), True),
+            ("0", (ValueNotExist, lambda x: float(x) >= 1), False),
+            ("hello", (ValueNotExist, lambda x: float(x) >= 1), False),
         ]
     )
     def test__not_passed(self, source, args, _EXPECTED):